+++ plugins/helper/iso8601time.go
@@ -1,4 +1,4 @@
-package core
+package helper
 
 import (
 	"fmt"
